Change the Orientation of the AcDisplayActivity to 'behind' (Closes #118)
This does mean that It'll start in the same orientation as the activity that was showing before.
From Android Developers: "The same orientation as the activity that's immediately beneath it in the activity stack."
I thought about use cases and this is the most appropriate approach. and it doesn't hurt current users because most of them will have their phone in Portrait orientation, and for the ones using landscape then okay, Lockscreen will start in landscape too, and the screen contents fit correctly.